backend/requirements.txt
# Core API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sse-starlette==2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Agno Agent dependencies
agno==1.5.0
openai>=1.101.0
sqlalchemy>=2.0.43
aiofiles>=24.1.0
chromadb>=1.0.20

# Financial data tools
yfinance>=0.2.65
yahoofinance>=0.0.2
openbb>=4.1.3

# Data processing
pandas>=2.0.0
plotly>=5.17.0

# HTTP client
httpx>=0.24.0
python-multipart>=0.0.6

# CLI and utilities
rich>=13.0.0
typer>=0.9.0
pyyaml>=6.0
docstring-parser>=0.15.0
gitpython>=3.1.0
tomli>=2.0.0

# Streamlit (if needed for hybrid deployment)
streamlit>=1.28.0

